Moved functions from index.php to a functions.php file and required them in. Added logic to the dropdown so it displays the selected filter after submit.
